[FIX #66] Migrate a few deprecated methods to reduce build warnings
* [FIX, WARN] Update Mode installation alert to no longer use deprecated +alertWithMessageText:defaultButton:alternateButton:otherButton:informativeTextWithFormat:.

Reviewed by @tolmasky.

* [FIX, WARN] Update info Mode installation NSAlert to no longer use deprecated +alertWithMessageText:defaultButton:alternateButton:otherButton:informativeTextWithFormat:.

Reviewed by @tolmasky.

* [FIX, WARN] Update -closeAllTabs in PlainTextWindowController to no longer use deprecated -beginSheetModalForWindow:modalDelegate:didEndSelector:contextInfo:.

Reviewed by @tolmasky.

* [FIX, WARN] Fix some conversion issues (mainly switching int to NSInteger and unsigned to NSUInteger).

Reviewed by @tolmasky.

* [FIX, WARN] Update -testValidityOfReplacementAndReportToUserForContext: in FindReplaceController to no longer use deprecated -beginSheetModalForWindow:modalDelegate:didEndSelector:contextInfo:.

Reviewed by @tolmasky.

* [FIX, WARN] Replace instances of deprecated -insertText: with insertText:replacementRange:.

Reviewed by @tolmasky.

* [FIX, WARN] Update AppController -checkUserModesForUpdateAfterVersionBump to no longer use deprecated +alertWithMessageText:defaultButton:alternateButton:otherButton:informativeTextWithFormat:.


* [FIX, WARN] Update AppController -reportAppleScriptError: to no longer use deprecated -beginSheetModalForWindow:modalDelegate:didEndSelector:contextInfo:.

Contributed by @tolmasky.
Reviewed by @monkeydom.

Pull Request #67

+++ SubEthaEdit-Mac/Source/SyntaxDefinition.m
@@ -1037,7 +1037,7 @@ static NSString * const StateDictionaryUseAutocompleteFromModeKey      = @"useau
     }
 
 
-    int combinedStringLength = [combinedString length];
+    NSUInteger combinedStringLength = [combinedString length];
     if ((combinedStringLength>1)||(endString)) {
         if (endString) { // Any states except the default
             [combinedString appendString:[NSString stringWithFormat:@"(?<seeinternalgroup4242>%@)",endString]];
